[ODN-90] change to public abstract BigInteger? NumItems { get; }//Null means infinity

+++ opticsdotnet/opticsdotnet.Lib/Objects/PossibleValueSet/PossibleValueSet.cs
@@ -4,7 +4,7 @@ namespace opticsdotnet.Lib
 {
     public abstract class PossibleValueSet : ITemplateSpot
     {
-        public abstract int? NumItems { get; }//Null means infinity
+        public abstract BigInteger? NumItems { get; }//Null means infinity
         public abstract Type ItemType { get; }
     }
